% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get.blocks.R
\name{get.blocks}
\alias{get.blocks}
\title{Get data.frame with data on all US Census 2010 blocks (pop, lat/lon, etc.)}
\usage{
get.blocks(fields = c("fips", "pop", "lat", "lon", "area", "urban"),
  charfips = TRUE)
}
\arguments{
\item{fields}{Optional vector of character elements specifying which fields to return.}

\item{charfips}{Optional TRUE by default, specifies if FIPS should be converted to character class with any necessary leading zeroes,
which uses more RAM and takes much longer -- It can take 1-2 minutes for this function to return results unless charfips=FALSE.}
}
\value{
Returns a (large, >11 million rows) dataframe that has specified fields
  or by default these 6 columns: fips, pop, lat, lon, area, urban
}
\description{
Returns a large dataframe with one row per block. This helps assemble the desired fields for all 11m+ blocks, into a single data.frame.
}
\details{
Warning: It can take 1-2 minutes for this function to return results with default settings (i.e., unless charfips=FALSE is specified).
  The full blocks data.frame created by default uses approximately 1 GB of RAM.
  The blocks data.frame with just numeric fips and pop uses only about 133 MB and is
}
\examples{
# blocks <- get.blocks()
# blocks <- get.blocks(c('fips','pop'))
}
\seealso{
\code{\link{blocks.fips}}
}
\concept{
demographics
}

