% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/UScensus2010blocks-package.R
\docType{package}
\name{UScensus2010blocks-package}
\alias{UScensus2010blocks}
\alias{UScensus2010blocks-package}
\alias{blocks}
\title{Census Bureau 2010 block-level data (pop, area, urban, lat/lon) easily compiled into a single data.frame}
\description{
These datasets provide population count, size of block (area), latitude and longitude of internal point,
  whether the block is urban, for each US block, based on Census Bureau Census 2010 data.
  All States/DC are compiled into a single vector for each field (e.g. pop).
}
\examples{
\dontrun{
   blocks <- get.blocks(charfips=FALSE)
   by(1e6 * blocks$pop / blocks$area, INDICES=blocks$urban, FUN=mean)
 }
}
\references{
\code{\link{http://ejanalysis.github.io}} \cr
\code{\link{http://www.ejanalysis.com/}} \cr
}

